fix!: use standard pyproject.toml attributes (#28), fix namespaced packages (#27)
This allows generated projects to be built with tools other than Poetry,
such as `uv`. This has been tested with `poetry build` and `uv build`.

BREAKING CHANGE: Author name and email address are now optional arguments
and not populated by default.  However, this does not seem to be required by
Poetry or `uv`.